#145 #179: update tests
