+++ test/test_mesh_seq.py
@@ -4,17 +4,9 @@ Testing for the mesh sequence objects.
 
 import re
 import unittest
-from unittest.mock import patch
 
-from firedrake import (
-    Function,
-    FunctionSpace,
-    UnitCubeMesh,
-    UnitSquareMesh,
-    UnitTriangleMesh,
-)
+from firedrake import Function, FunctionSpace, UnitCubeMesh, UnitSquareMesh
 from parameterized import parameterized
-from pyadjoint.block_variable import BlockVariable
 
 from goalie.mesh_seq import MeshSeq
 from goalie.time_partition import TimeInterval, TimePartition
@@ -177,135 +169,3 @@ class TestStringFormatting(unittest.TestCase):
             "Mesh(VectorElement(FiniteElement('Lagrange', triangle, 1), dim=2), .*)])"
         )
         self.assertTrue(re.match(repr(mesh_seq), expected))
-
-
-class TestBlockLogic(unittest.TestCase):
-    """
-    Unit tests for :meth:`MeshSeq._dependency` and :meth:`MeshSeq._output`.
-    """
-
-    @staticmethod
-    def get_p0_spaces(mesh):
-        return {"field": FunctionSpace(mesh, "DG", 0)}
-
-    def setUp(self):
-        self.time_interval = TimeInterval(1.0, 0.5, "field")
-        self.mesh = UnitTriangleMesh()
-        self.mesh_seq = MeshSeq(
-            self.time_interval, self.mesh, get_function_spaces=self.get_p0_spaces
-        )
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_output_not_function(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        block_variable = BlockVariable(1)
-        solve_block._outputs = [block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._output("field", 0, solve_block)
-        msg = "Solve block for field 'field' on subinterval 0 has no outputs."
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_output_wrong_function_space(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        block_variable = BlockVariable(Function(FunctionSpace(self.mesh, "CG", 1)))
-        solve_block._outputs = [block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._output("field", 0, solve_block)
-        msg = "Solve block for field 'field' on subinterval 0 has no outputs."
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_output_wrong_name(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        function_space = FunctionSpace(self.mesh, "DG", 0)
-        block_variable = BlockVariable(Function(function_space, name="field2"))
-        solve_block._outputs = [block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._output("field", 0, solve_block)
-        msg = "Solve block for field 'field' on subinterval 0 has no outputs."
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_output_valid(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        function_space = FunctionSpace(self.mesh, "DG", 0)
-        block_variable = BlockVariable(Function(function_space, name="field"))
-        solve_block._outputs = [block_variable]
-        self.assertIsNotNone(self.mesh_seq._output("field", 0, solve_block))
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_output_multiple_valid_error(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        function_space = FunctionSpace(self.mesh, "DG", 0)
-        block_variable = BlockVariable(Function(function_space, name="field"))
-        solve_block._outputs = [block_variable, block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._output("field", 0, solve_block)
-        msg = (
-            "Cannot determine a unique output index for the solution associated with"
-            " field 'field' out of 2 candidates."
-        )
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_dependency_not_function(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        block_variable = BlockVariable(1)
-        solve_block._dependencies = [block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._dependency("field", 0, solve_block)
-        msg = "Solve block for field 'field' on subinterval 0 has no dependencies."
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_dependency_wrong_function_space(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        block_variable = BlockVariable(Function(FunctionSpace(self.mesh, "CG", 1)))
-        solve_block._dependencies = [block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._dependency("field", 0, solve_block)
-        msg = "Solve block for field 'field' on subinterval 0 has no dependencies."
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_dependency_wrong_name(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        function_space = FunctionSpace(self.mesh, "DG", 0)
-        block_variable = BlockVariable(Function(function_space, name="field_new"))
-        solve_block._dependencies = [block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._dependency("field", 0, solve_block)
-        msg = "Solve block for field 'field' on subinterval 0 has no dependencies."
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_dependency_valid(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        function_space = FunctionSpace(self.mesh, "DG", 0)
-        block_variable = BlockVariable(Function(function_space, name="field_old"))
-        solve_block._dependencies = [block_variable]
-        self.assertIsNotNone(self.mesh_seq._dependency("field", 0, solve_block))
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_dependency_multiple_valid_error(self, MockSolveBlock):
-        solve_block = MockSolveBlock()
-        function_space = FunctionSpace(self.mesh, "DG", 0)
-        block_variable = BlockVariable(Function(function_space, name="field_old"))
-        solve_block._dependencies = [block_variable, block_variable]
-        with self.assertRaises(AttributeError) as cm:
-            self.mesh_seq._dependency("field", 0, solve_block)
-        msg = (
-            "Cannot determine a unique dependency index for the lagged solution"
-            " associated with field 'field' out of 2 candidates."
-        )
-        self.assertEqual(str(cm.exception), msg)
-
-    @patch("firedrake.adjoint_utils.blocks.solving.GenericSolveBlock")
-    def test_dependency_steady(self, MockSolveBlock):
-        time_interval = TimeInterval(1.0, 0.5, "field", field_types="steady")
-        mesh_seq = MeshSeq(
-            time_interval, self.mesh, get_function_spaces=self.get_p0_spaces
-        )
-        solve_block = MockSolveBlock()
-        self.assertIsNone(mesh_seq._dependency("field", 0, solve_block))
