+++ test_adjoint/examples/point_discharge2d.py
@@ -56,8 +56,8 @@ def get_form(self):
     Advection-diffusion with SUPG stabilisation.
     """
 
-    def form(i, sols):
-        c, c_ = sols["tracer_2d"]
+    def form(i):
+        _, c = self.fields["tracer_2d"]
         fs = self.function_spaces["tracer_2d"][i]
         R = FunctionSpace(self[i], "R", 0)
         D = Function(R).assign(0.1)
@@ -91,15 +91,14 @@ def get_solver(self):
     solved using a direct method.
     """
 
-    def solver(i, ic):
+    def solver(i):
         fs = self.function_spaces["tracer_2d"][i]
 
-        # Ensure dependence on initial condition
-        c = Function(fs, name="tracer_2d_old")
-        c.assign(ic["tracer_2d"])
+        _, c = self.fields["tracer_2d"]
+        self.fields["tracer_2d"] = (c, c)
 
         # Setup variational problem
-        F = self.form(i, {"tracer_2d": (c, c)})["tracer_2d"]
+        F = self.form(i)["tracer_2d"]
 
         # Zero Dirichlet condition on the left-hand (inlet) boundary
         bc = DirichletBC(fs, 0, 1)
@@ -113,12 +112,12 @@ def get_solver(self):
             "pc_factor_mat_solver_type": "mumps",
         }
         solve(F == 0, c, bcs=bc, solver_parameters=sp, ad_block_tag="tracer_2d")
-        return {"tracer_2d": c}
+        yield
 
     return solver
 
 
-def get_qoi(self, sol, i):
+def get_qoi(self, i):
     """
     Quantity of interest which integrates
     the tracer concentration over an offset
@@ -126,7 +125,7 @@ def get_qoi(self, sol, i):
     """
 
     def steady_qoi():
-        c = sol["tracer_2d"]
+        c = self.fields["tracer_2d"][0]
         x, y = SpatialCoordinate(self[i])
         kernel = conditional((x - rec_x) ** 2 + (y - rec_y) ** 2 < rec_r**2, 1, 0)
         area = assemble(kernel * dx)
