+++ test_adjoint/test_adjoint.py
@@ -42,7 +42,7 @@ class TestAdjointMeshSeqGeneric(unittest.TestCase):
     def test_get_qoi_notimplemented_error(self):
         mesh_seq = AdjointMeshSeq(self.time_interval, self.meshes, qoi_type="end_time")
         with self.assertRaises(NotImplementedError) as cm:
-            mesh_seq.get_qoi({}, 0)
+            mesh_seq.get_qoi(0)
         msg = "'get_qoi' is not implemented."
         self.assertEqual(str(cm.exception), msg)
 
@@ -128,8 +128,12 @@ def test_adjoint_same_mesh(problem, qoi_type, debug=False):
     tape.clear_tape()
     ic = mesh_seq.initial_condition
     controls = [pyadjoint.Control(value) for key, value in ic.items()]
-    sols = mesh_seq.solver(0, ic)
-    qoi = mesh_seq.get_qoi(sols, 0)
+    mesh_seq._reinitialise_fields(ic)
+    solver = mesh_seq.solver(0)
+    for i in range(len(mesh_seq)):
+        for _ in range(mesh_seq.time_partition.num_timesteps_per_subinterval[i]):
+            next(solver)
+    qoi = mesh_seq.get_qoi(0)
     J = mesh_seq.J if qoi_type == "time_integrated" else qoi()
     m = pyadjoint.enlisting.Enlist(controls)
     assert pyadjoint.annotate_tape()
@@ -146,6 +150,7 @@ def test_adjoint_same_mesh(problem, qoi_type, debug=False):
     adj_values_expected = {}
     for field, fs in mesh_seq._fs.items():
         solve_blocks = mesh_seq.get_solve_blocks(field, 0)
+        firedrake.PETSc.Sys.Print(solve_blocks)
         adj_sols_expected[field] = solve_blocks[first_export_idx].adj_sol.copy(
             deepcopy=True
         )
